Refactor Todo App structure by creating separate components for task input, task list, filtering, and check all functionality
