+++ src/component/TaskList.js
@@ -1,0 +1,84 @@
+import React, { useState } from 'react';
+
+function TaskListItem({ task, remove_task, check_single_task, update_single_task }) {
+
+    const [focused, setFocused] = useState(false);
+    const [name, setName] = useState(task.name)
+
+    return (
+        <li className="todo-item-container">
+            <div className="todo-item">
+                <input type="checkbox" checked={task.isCompleted} onChange={() => check_single_task(task.id)} />
+                {focused ?
+                    (<input
+                        type="text"
+                        className="todo-input"
+                        value={name}
+                        onBlur={() => {
+                            update_single_task(name, task.id)
+                            setFocused(false);
+                        }}
+                        onChange={(e) => setName(e.target.value)}
+                    />
+                    ) :
+                    (<span
+                        onClick={() => setFocused(true)}
+                        className={`todo-item-label ${task.isCompleted ? "completed-todo-item" : ""}`}>
+                        {name}
+                    </span>
+                    )
+                }
+                {/* <input type="text" className="todo-item-input" value="Finish React Series" /> */}
+            </div>
+            <button className="x-button" onClick={() => remove_task(task.id)}>
+                <svg
+                    className="x-button-icon"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke="currentColor"
+                >
+                    <path
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                        strokeWidth={2}
+                        d="M6 18L18 6M6 6l12 12"
+                    />
+                </svg>
+            </button>
+        </li>
+    )
+
+}
+
+
+export default function TaskList({taskList, renderFiliter, setTaskList}) {
+
+      function remove_task(delete_task_id) {
+        setTaskList(taskList.filter(task => task.id !== delete_task_id));
+      }
+    
+      function check_single_task(check_task_id) {
+        setTaskList(taskList.map(value => (value.id === check_task_id) ? { ...value, isCompleted: !value.isCompleted } : value
+        ));
+      };
+    
+      function update_single_task(updated_value, update_task_id) {
+        setTaskList(taskList.map(value => (value.id === update_task_id) ? { ...value, name: updated_value } : value
+        ));
+      }
+
+    return (
+    <ul className="todo-list">
+        {
+            taskList.filter(task => {
+                if (renderFiliter === 'All') return true;
+                else if (renderFiliter === 'Active' && task.isCompleted === false) return true;
+                else if (renderFiliter === 'Completed' && task.isCompleted === true) return true;
+                return false;
+            }).map((task, index) => {
+                return <TaskListItem key={task.id} task={task} remove_task={remove_task} check_single_task={check_single_task} update_single_task={update_single_task} />
+            })
+        }
+    </ul>
+    )
+}
